requirements.txt
streamlit
pandas
openpyxl
pycryptodome
streamlit-option-menu

